создаёт дополнительные модули
