+++ src/const.js
@@ -1,0 +1,19 @@
+export const TRIP_EVENT_TYPES = [
+  'taxi',
+  'bus',
+  'train',
+  'ship',
+  'drive',
+  'flight',
+  'check-in',
+  'sightseeing',
+  'restaurant',
+];
+
+export const TRIP_CITIES = [
+  'Amsterdam',
+  'Berlin',
+  'Chamonix',
+  'Dresden',
+  'Edinburgh',
+];
