+++ src/view/trip-event-view.js
@@ -28,6 +28,8 @@ export const createTripEvent = ({basePrice, dateFrom, dateTo, destination, isFav
     return '';
   };
 
+  const getFavoriteClass = () => isFavorite ? ' event__favorite-btn--active' : '';
+
   return `<li class="trip-events__item">
     <div class="event">
       <time class="event__date" datetime="${startTime.format('YYYY-MM-DD')}">${startTime.format('MMM DD')}</time>
@@ -50,7 +52,7 @@ export const createTripEvent = ({basePrice, dateFrom, dateTo, destination, isFav
       <ul class="event__selected-offers">
         ${getOfferNodes()}
       </ul>
-      <button class="event__favorite-btn${isFavorite ? ' event__favorite-btn--active' : ''}" type="button">
+      <button class="event__favorite-btn${getFavoriteClass()}" type="button">
         <span class="visually-hidden">Add to favorite</span>
         <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
           <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
